refactor(profiles): merge all members into a file
