+++ src/components/Committee.astro
@@ -1,31 +1,6 @@
 ---
-import type { ChairsInterface } from "./Profile/chairs/types";
-import challengeChairs from "./Profile/chairs/challenge/profiles";
-import financialChairs from "./Profile/chairs/financial/profiles";
-import generalChairs from "./Profile/chairs/general/profiles";
-import honoraryChairs from "./Profile/chairs/honarary/profiles";
-import industrialChairs from "./Profile/chairs/industrial/profiles";
-import programChairs from "./Profile/chairs/program/profiles";
-import publicationChairs from "./Profile/chairs/publication/profiles";
-import publicityChairs from "./Profile/chairs/publicity/profiles";
-import tutorialChairs from "./Profile/chairs/tutorial/profiles";
-import websiteAndRestrationChairs from "./Profile/chairs/websiteAndRegistration/profiles";
-import workshopChairs from "./Profile/chairs/workshop/profiles";
 import Profile from "./Profile/Profile.astro";
-
-const committee: ChairsInterface[] = [
-  generalChairs,
-  honoraryChairs,
-  programChairs,
-  publicationChairs,
-  financialChairs,
-  publicityChairs,
-  websiteAndRestrationChairs,
-  workshopChairs,
-  tutorialChairs,
-  industrialChairs,
-  challengeChairs,
-];
+import committee from "./Profile/profiles";
 ---
 
 {
