--- src/components/Profile/chairs/challenge/profiles.ts
@@ -1,19 +1,0 @@
-import type { ChairsInterface } from "../types";
-
-const challengeChairs: ChairsInterface = {
-  roll: "Challenge Chairs",
-  profiles: [
-    {
-      name: "Zheng Tang",
-      description: "Nvidia Metropolis, USA",
-      profileImage: "committee/zheng-tang.png",
-    },
-    {
-      name: "Weiyao Lin",
-      description: "SJTU, China",
-      profileImage: "committee/weiyao-lin.png",
-    },
-  ],
-};
-
-export default challengeChairs;
