--- src/components/Profile/chairs/financial/profiles.ts
@@ -1,19 +1,0 @@
-import type { ChairsInterface } from "../types";
-
-const financialChairs: ChairsInterface = {
-  roll: "Financial Chairs",
-  profiles: [
-    {
-      name: "Deokwoo Lee",
-      description: "Keimyung University, S.Korea",
-      profileImage: "committee/deokwoo-lee.png",
-    },
-    {
-      name: "Jun-Wei Hsieh",
-      description: "NYCU, Taiwan",
-      profileImage: "committee/jun-wei-hsieh.png",
-    },
-  ],
-};
-
-export default financialChairs;
