--- src/components/Profile/chairs/general/profiles.ts
@@ -1,19 +1,0 @@
-import type { ChairsInterface } from "../types";
-
-const generalChairs: ChairsInterface = {
-  roll: "General Chairs",
-  profiles: [
-    {
-      name: "Jeng-Neung Hwang",
-      description: "University of Washington, USA",
-      profileImage: "committee/jengneung-hwang.png",
-    },
-    {
-      name: "Michael S. Ryoo",
-      description: "Stony Brook University & Google Brain, USA",
-      profileImage: "committee/michael-ryoo.png",
-    },
-  ],
-};
-
-export default generalChairs;
