--- src/components/Profile/chairs/honarary/profiles.ts
@@ -1,19 +1,0 @@
-import type { ChairsInterface } from "../types";
-
-const honoraryChairs: ChairsInterface = {
-  roll: "Honorary Chairs",
-  profiles: [
-    {
-      name: "Dong Seog Han",
-      description: "KNU, S.Korea",
-      profileImage: "committee/dongseog-han.png",
-    },
-    {
-      name: "Eun Chang Choi",
-      description: "ETRI, S.Korea",
-      profileImage: "committee/eunchang-choi.png",
-    },
-  ],
-};
-
-export default honoraryChairs;
