--- src/components/Profile/chairs/industrial/profiles.ts
@@ -1,19 +1,0 @@
-import type { ChairsInterface } from "../types";
-
-const industrialChairs: ChairsInterface = {
-  roll: "Industrial Chairs",
-  profiles: [
-    {
-      name: "Peter Tu",
-      description: "GE Global Research, USA",
-      profileImage: "committee/peter-tu.png",
-    },
-    {
-      name: "Lipeng Ke",
-      description: "Amazon Lab126, USA",
-      profileImage: "committee/lipeng-ke.png",
-    },
-  ],
-};
-
-export default industrialChairs;
