--- src/components/Profile/chairs/program/profiles.ts
@@ -1,29 +1,0 @@
-import type { ChairsInterface } from "../types";
-
-const programChairs: ChairsInterface = {
-  roll: "Program Chairs",
-  profiles: [
-    {
-      name: "Kwang-Ju Kim",
-      description: "ETRI, S.Korea",
-      profileImage: "committee/kwangju-kim.png",
-    },
-    {
-      name: "Jong Taek Lee",
-      description: "KNU, S.Korea",
-      profileImage: "committee/jongtaek-lee.png",
-    },
-    {
-      name: "Byoung Chul Ko",
-      description: "Keimyung University, S.Korea",
-      profileImage: "committee/byoungchul-ko.png",
-    },
-    {
-      name: "Arslan Basharat",
-      description: "Kitware Inc, USA",
-      profileImage: "committee/arslan-basharat.png",
-    },
-  ],
-};
-
-export default programChairs;
