--- src/components/Profile/chairs/publication/profiles.ts
@@ -1,14 +1,0 @@
-import type { ChairsInterface } from "../types";
-
-const publicationChairs: ChairsInterface = {
-  roll: "Publication Chairs",
-  profiles: [
-    {
-      name: "Srirangaraj Setlur",
-      description: "University of Buffalo, USA",
-      profileImage: "committee/srirangaraj-setlur.png",
-    },
-  ],
-};
-
-export default publicationChairs;
