--- src/components/Profile/chairs/publicity/profiles.ts
@@ -1,24 +1,0 @@
-import type { ChairsInterface } from "../types";
-
-const publicityChairs: ChairsInterface = {
-  roll: "Publicity Chairs",
-  profiles: [
-    {
-      name: "Jiaying Liu",
-      description: "Peking University, China",
-      profileImage: "committee/jiaying-liu.png",
-    },
-    {
-      name: "Wei Qi Yan",
-      description: "AUT, New Zealand",
-      profileImage: "committee/wei-qi-yan.png",
-    },
-    {
-      name: "Ruben Tolosana",
-      description: "Universidad Autonoma de Madrid, Spain",
-      profileImage: "committee/ruben-tolosana.png",
-    },
-  ],
-};
-
-export default publicityChairs;
