--- src/components/Profile/chairs/tutorial/profiles.ts
@@ -1,19 +1,0 @@
-import type { ChairsInterface } from "../types";
-
-const tutorialChairs: ChairsInterface = {
-  roll: "Tutorial Chairs",
-  profiles: [
-    {
-      name: "ChangIck Kim",
-      description: "KAIST, S.Korea",
-      profileImage: "committee/changick-kim.png",
-    },
-    {
-      name: "Gaoang Wang",
-      description: "ZJU-UIUC Institute, China",
-      profileImage: "committee/gaoang-wang.png",
-    },
-  ],
-};
-
-export default tutorialChairs;
