--- src/components/Profile/chairs/types.ts
@@ -1,10 +1,0 @@
-export interface ProfileInterface {
-  name: string;
-  description: string;
-  profileImage: string;
-}
-
-export interface ChairsInterface {
-  roll: string;
-  profiles: ProfileInterface[];
-}
