--- src/components/Profile/chairs/websiteAndRegistration/profiles.ts
@@ -1,19 +1,0 @@
-import type { ChairsInterface } from "../types";
-
-const websiteAndRestrationChairs: ChairsInterface = {
-  roll: "Website and Registration Chairs",
-  profiles: [
-    {
-      name: "Heeseok Oh",
-      description: "Hansung University, S.Korea",
-      profileImage: "committee/heeseok-oh.png",
-    },
-    {
-      name: "Yiwei Fu",
-      description: "GE Research, USA",
-      profileImage: "committee/yiwei-fu.png",
-    },
-  ],
-};
-
-export default websiteAndRestrationChairs;
