--- src/components/Profile/chairs/workshop/profiles.ts
@@ -1,27 +1,0 @@
-import type { ChairsInterface } from "../types";
-/**
- *
- * Hyun Jong Yang
- * POSTECH, S. Korea
- *
- * Federica Battisti
- * University of Padua, Italy
- */
-
-const workshopChairs: ChairsInterface = {
-  roll: "Workshop Chairs",
-  profiles: [
-    {
-      name: "Hyunjong Yang",
-      description: "POSTECH, S. Korea",
-      profileImage: "committee/hyunjong-yang.png",
-    },
-    {
-      name: "Federica Battisti",
-      description: "University of Padua, Italy",
-      profileImage: "committee/federica-battisti.png",
-    },
-  ],
-};
-
-export default workshopChairs;
